Crate pallet_preimage

Preimage Pallet

Config

Call

Overview

The Preimage pallet allows for the users and the runtime to store the preimage of a hash on chain. This can be used by other pallets for storing and managing large byte-blobs.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

migration

Storage migrations for the preimage pallet.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_preimage

Enums

OldRequestStatus

A type to note whether a preimage is owned by a user or the system.

RequestStatus

A type to note whether a preimage is owned by a user or the system.

Constants

MAX_HASH_UPGRADE_BULK_COUNT

Hard-limit on the number of hashes that can be passed to ensure_updated.

MAX_SIZE

Maximum size of preimage we can store is 4mb.

Type Aliases

BalanceOf

TicketOf