Trait pallet_preimage::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;
    type Currency: ReservableCurrency<Self::AccountId>;
    type ManagerOrigin: EnsureOrigin<Self::Origin>;
    type MaxSize: Get<u32>;
    type BaseDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type ByteDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type WeightInfo: WeightInfo

The Weight information for this pallet.

type Currency: ReservableCurrency<Self::AccountId>

Currency type for this pallet.

type ManagerOrigin: EnsureOrigin<Self::Origin>

An origin that can request a preimage be placed on-chain without a deposit or fee, or manage existing preimages.

type MaxSize: Get<u32>

Max size allowed for a preimage.

type BaseDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The base deposit for placing a preimage on chain.

type ByteDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The per-byte deposit for placing a preimage on chain.